move permissions files to dedicated folders
