+++ src/modules/permissions/controllers/permission.controller.ts
@@ -1,13 +1,15 @@
 import { Controller, Get } from '@nestjs/common'
-import { ApiOAuth2, ApiTags } from '@nestjs/swagger'
-import { Permission } from '../permission.enum.js'
+import { ApiBearerAuth, ApiResponse, ApiTags } from '@nestjs/swagger'
+import { Permission } from '../enums/permission.enum.js'
+import { getPermissionsResponse } from '../docs/permission-response.docs.js'
 
 @ApiTags('Permissions')
 @Controller('permissions')
-@ApiOAuth2([])
+@ApiBearerAuth()
 export class PermissionController {
-  @Get()
-  getPermissions (): Permission[] {
+  @Get('/')
+  @ApiResponse(getPermissionsResponse)
+  public getPermissions (): Permission[] {
     return Object.values(Permission)
   }
 }
