+++ src/modules/permissions/decorators/permissions.decorator.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import { SetMetadata } from '@nestjs/common'
-import { type Permission } from './permission.enum.js'
+import { type Permission } from '../enums/permission.enum.js'
 
 export const PERMISSIONS_KEY = 'permissions'
 export const Permissions = (...roles: Permission[]) => SetMetadata(PERMISSIONS_KEY, roles)
