+++ src/modules/permissions/docs/permission-response.docs.ts
@@ -1,0 +1,15 @@
+import { type ApiResponseOptions } from '@nestjs/swagger'
+import { Permission } from '../enums/permission.enum.js'
+
+export const getPermissionsResponse: ApiResponseOptions = {
+  status: 200,
+  description: 'The permissions have been successfully received.',
+  schema: {
+    type: 'array',
+    items: {
+      type: 'enum',
+      enum: Object.values(Permission)
+    },
+    example: Object.values(Permission)
+  }
+}
