+++ src/modules/permissions/enums/permission.enum.ts
